Tidyup prep work (refactor functions into misc)
